+++ errors/__tests__/errors.test.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const errors = require('..');
+
+describe('errors', () => {
+    it('needs tests');
+});