+++ packages/core/lib/constants.js
@@ -17,8 +17,9 @@ module.exports.lifecycleHook = [
 
 module.exports.logLevel = {
   trace: 'trace',
-  info: 'info',
   debug: 'debug',
+  info: 'info',
   warn: 'warn',
+  error: 'error',
   fatal: 'fatal'
 }