+++ packages/core/lib/registry/service.js
@@ -1,4 +1,3 @@
-const { cloneDeep } = require('lodash')
 const { mergeSchemas } = require('../utils/options')
 const { wrapInArray, isFunction, clone, wrapHandler, promisify, isObject } = require('@weave-js/utils')
 
@@ -90,7 +89,7 @@ const createService = (broker, middlewareHandler, addLocalService, registerLocal
         action = wrapHandler(action)
       }
 
-      const innerAction = createActionHandler(cloneDeep(action), name)
+      const innerAction = createActionHandler(clone(action), name)
       registryItem.actions[innerAction.name] = innerAction
 
       const wrappedAction = middlewareHandler.wrapHandler('localAction', innerAction.handler, innerAction)
@@ -230,16 +229,15 @@ const createService = (broker, middlewareHandler, addLocalService, registerLocal
 
   function applyMixins (schema) {
     const mixins = wrapInArray(schema.mixins)
-
     if (mixins.length > 0) {
       const mixedSchema = Array
         .from(mixins)
         .reverse()
         .reduce((s, mixin) => {
+          if (mixin.mixins) {
+            mixin = applyMixins(mixin)
+          }
           for (var key in mixin) {
-            if (mixin.mixins) {
-              mixin = applyMixins(mixin)
-            }
             // bind scope for life cycle hooks
             if (lifecycleHook.includes(key)) {
               mixin[key] = mixin[key].bind(self)